cleaned warnings in logger_bbob and large_scale_transformation

+++ code-experiments/src/logger_bbob.c
@@ -322,8 +322,8 @@ static void logger_bbob_initialize(logger_bbob_t *logger, coco_problem_t *inner_
   char indexFile_prefix[10] = "bbobexp"; /* TODO (minor): make the prefix bbobexp a parameter that the user can modify */
   size_t str_length_funId, str_length_dim;
   
-  str_length_funId = (size_t) bbob2009_fmax(1, ceil(log10(coco_problem_get_suite_dep_function(inner_problem))));
-  str_length_dim = (size_t) bbob2009_fmax(1, ceil(log10(inner_problem->number_of_variables)));
+  str_length_funId = (size_t) bbob2009_fmax(1, ceil(log10((double) coco_problem_get_suite_dep_function(inner_problem))));
+  str_length_dim = (size_t) bbob2009_fmax(1, ceil(log10((double) inner_problem->number_of_variables)));
   tmpc_funId = (char *) coco_allocate_memory(str_length_funId *  sizeof(char));
   tmpc_dim = (char *) coco_allocate_memory(str_length_dim *  sizeof(char));
